b-components: changing the name of the components
