+++ src/app/components/Header/Header.scss
@@ -1,0 +1,36 @@
+// SCSS base
+@import "../../App.scss";
+
+.header {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: flex-start;
+  height: 100%;
+
+  .headerLogo {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    img {
+      width: 2.8em;
+    }
+  }
+
+  .navbarTitle {
+    font-size: 18px;
+    font-weight: bold;
+    line-height: 26px;
+
+    display: flex;
+    align-items: center;
+    letter-spacing: 0.15em;
+    transform: rotate(270deg);
+
+    padding-left: 10px;
+    margin-left: -25px;
+
+    cursor: pointer;
+  }
+}
