+++ src/app/components/Header/Header.tsx
@@ -1,0 +1,22 @@
+import * as React from "react";
+import "./Header.scss";
+
+import Logo from "../../../assets/logo.svg";
+
+export const Header = ({ setNavbarStatus }: any): JSX.Element => {
+  return (
+    <nav className="header">
+      <div className="headerLogo">
+        <img src={Logo} alt="logo" />
+      </div>
+      <p
+        className="navbarTitle"
+        onClick={() => {
+          setNavbarStatus(true);
+        }}
+      >
+        menu
+      </p>
+    </nav>
+  );
+};
